FIX: DiscountPlicy 필드 수정

+++ src/main/java/com/ayuconpon/coupon/domain/value/DiscountPolicy.java
@@ -3,6 +3,8 @@ package com.ayuconpon.coupon.domain.value;
 import com.ayuconpon.common.Money;
 import com.ayuconpon.common.MoneyConverter;
 import jakarta.persistence.*;
+import jakarta.validation.constraints.DecimalMax;
+import jakarta.validation.constraints.DecimalMin;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
@@ -14,19 +16,19 @@ public class DiscountPolicy {
     @Column(name = "discount_type")
     private DiscountType discountType;
     @Column(name = "discount_rate")
-    private Long discountRate;
+    private Float discountRate;
     @Convert(converter = MoneyConverter.class)
     @Column(name = "discount_value")
     private Money discountValue;
 
-    private DiscountPolicy(DiscountType discountType, Long discountRate, Money discountValue) {
+    private DiscountPolicy(DiscountType discountType, Float discountRate, Money discountValue) {
         this.discountType = discountType;
         this.discountRate = discountRate;
         this.discountValue = discountValue;
     }
 
-    public static DiscountPolicy of(DiscountType discountType, Long discountRate, Money discountValue) {
-        return new DiscountPolicy(discountType,discountRate, discountValue);
+    public static DiscountPolicy of(DiscountType discountType, Float discountRate, Money discountValue) {
+        return new DiscountPolicy(discountType, discountRate, discountValue);
     }
 
 }
